# peerstore: fix addressbook benchmark timing